requirements.txt
pywin32==311
google-genai==1.64.0
python-dotenv==1.2.1
